# Refs #28459 -- Used default date converter on SQLite for better performance.