https://code.djangoproject.com/ticket/28459

See https://docs.python.org/3/library/sqlite3.html#default-adapters-and-converters.

Before:
```
%timeit for x in TestModel.objects.values_list('date'): pass
83.3 ms ± 829 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
After:
```
%timeit for x in TestModel.objects.values_list('date'): pass
40.7 ms ± 613 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```